requirements.txt
selenium
beautifulsoup4
pandas
python-dotenv
